src/Dashboard/db_connector/pyproject.toml
[project]
name = "multi-db-connector"
version = "0.1.0"
description = "Generic database connector package for Python projects"
authors = [
    "David Camilo Corrales <David-Camilo.Corrales-Munoz@inrae.fr>",
    "Carlos Alberto Suarez Muñoz <csuarez@gmail.com>"
]
dependencies = [
    "python-dotenv",
    "influxdb-client",
    "psycopg2-binary",
    "pandas",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"